% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truss0.R
\name{truss0}
\alias{truss0}
\title{Cluster by k-truss: Deprecated version.}
\usage{
truss0(g, k)
}
\arguments{
\item{k}{The level of support, requiring that every edge in subgraph is in
at least (k-2) triangles (J. Cohen 2008). \code{k} must be at least 3.}

\item{graph}{An undirected graph (must be igraph).}
}
\value{
A subgraph containing all k-trusses in the graph for a given level of support \code{k}.
}
\description{
Compute the k-truss of an undirected graph (deprecated).
}
\examples{

require(igraph)
set.seed(1)
g <- sample_gnm(6, 9)
E(g)$eid <- seq(ecount(g))
g_truss <- truss(g, k = 3)

# Plot original graph.
fixed_layout <- layout_with_lgl(g)
plot(g, layout = fixed_layout)  # original

# Plot graph with truss clustering.
E(g)$weight <- 1
E(g)$color <- "gray"
E(g)[E(g_truss)$eid]$color <- "blue"
E(g)[E(g_truss)$eid]$weight <- 3
plot(g, layout = fixed_layout, edge.width = E(g)$weight)  # with truss
}
\keyword{graphs}

