% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truss.R
\name{truss_edge}
\alias{truss_edge}
\title{Cluster by k-truss}
\usage{
truss_edge(g)
}
\arguments{
\item{graph}{An undirected graph (must be igraph).}
}
\value{
Vector indicating the maximal \code{k}-truss that each edge belongs to (NA if none).
}
\description{
Compute maximal \code{k}-truss each edge belongs to.
}
\examples{

require(igraph)
set.seed(1)
g <- sample_gnm(6, 9)
E(g)$truss <- truss_edge(g)

# Plot original graph.
fixed_layout <- layout_with_lgl(g)
plot(g, layout = fixed_layout)  # original

# Plot graph with truss clustering.
E(g)$weight <- 1
E(g)$color <- "gray"
E(g)[which(E(g)$truss <= 3)]$color <- "blue"
E(g)[which(E(g)$truss <= 3)]$weight <- 3
plot(g, layout = fixed_layout, edge.width = E(g)$weight)  # with truss
}
\keyword{graphs}
